backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
cachetools
twilio
